docs: update README to include dependency injection section and usage examples; add DTOs for user operations
