+++ src/modules/users/repositories/IUsersRepository.ts
@@ -1,21 +1,5 @@
 import { User } from '../entities/User'
-
-interface ICreateUserDTO {
-  id?: string
-  name: string
-  email: string
-  company: string
-  password: string
-}
-
-interface PublicUserInfoDTO {
-  id: string
-  name: string
-  email: string
-  company: string
-  created_at: Date
-  updated_at: Date
-}
+import { ICreateUserDTO, PublicUserInfoDTO } from '../dtos'
 
 interface IUsersRepository {
   createUser({ name, email, company, password }: ICreateUserDTO): Promise<User>
@@ -26,4 +10,4 @@ interface IUsersRepository {
   updateUser({ id, name, email, company, password }: ICreateUserDTO): Promise<any>
 }
 
-export { ICreateUserDTO, IUsersRepository, PublicUserInfoDTO }
+export { IUsersRepository }
