+++ src/modules/users/useCases/updateUser/UpdateUserUseCase.ts
@@ -2,14 +2,7 @@ import { inject, injectable } from 'tsyringe'
 import { AppError } from '../../../../errors/AppError'
 import { User } from '../../entities/User'
 import { IUsersRepository } from '../../repositories/IUsersRepository'
-
-interface IRequest {
-  id: string
-  name: string
-  email: string
-  company: string
-  password: string
-}
+import { IUpdateUserDTO } from '../../dtos'
 
 @injectable()
 class UpdateUserUseCase {
@@ -18,7 +11,7 @@ class UpdateUserUseCase {
     private usersRepository: IUsersRepository
   ) {}
 
-  async execute({ id, name, email, company, password }: IRequest): Promise<User> {
+  async execute({ id, name, email, company, password }: IUpdateUserDTO): Promise<User> {
     const userExists = await this.usersRepository.findUserById(id)
 
     if (!userExists) {
